backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
motor==3.1.1
multidict==6.0.4
orjson==3.8.9
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
twilio==8.0.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
yarl==1.8.2
